requirements.txt
selenium==4.1.3
pytest==7.1.1
autopep8==1.6.0
flake8==4.0.1
faker==13.4.0
allure-pytest==2.9.45
